SLqPCR-package {SLqPCR}	R Documentation

Functions for analysis of real-time quantitative PCR data at SIRS-Lab GmbH

Description

Functions for analysis of real-time quantitative PCR data at SIRS-Lab GmbH

Details

Package: 	 SLqPCR

Type: 	 Package

Version: 	 1.0.0

Date: 	 2007-01-02

Depends: 	 R(>= 2.4.0), stats, RColorBrewer

License: 	 GPL (version 2 or later)

require(SLqPCR)

Author(s)

Dr. Matthias Kohl (SIRS-Lab GmbH) http://www.sirs-lab.com

Maintainer: Dr. Matthias Kohl kohl@sirs-lab.com

References

Jo Vandesompele, Katleen De Preter, Filip Pattyn et al. (2002). Accurate normalization of real-time quantitative RT-PCR data by geometric averaging of multiple internal control genes. Genome Biology 2002. 3(7):research0034.1-0034.11. http://genomebiology.com/2002/3/7/research/0034/

[Package SLqPCR version 1.60.0 Index]