geomMean {SLqPCR}	R Documentation

Geometric Mean

Description

Computation of the geometric mean.

Usage

geomMean(x, na.rm = FALSE)

Arguments

x
numeric vector of non-negative Reals

na.rm
a logical value indicating whether NA values should be stripped before the computation proceeds.

Details

The computation of the geometric mean is done via prod(x)^(1/length(x)).

Value

geometric mean

Author(s)

Dr. Matthias Kohl (SIRS-Lab GmbH) kohl@sirs-lab.com

[Package SLqPCR version 1.60.0 Index]