normPCR {SLqPCR}	R Documentation

Normalization of real-time quantitative RT-PCR data

Description

This function can be used to normalize real-time quantitative RT-PCR data.

Usage

normPCR(relData, HKs, method = "Vandesompele", na.rm = FALSE)

Arguments

relData
matrix or data.frame containing relative quantities (genes in columns)

HKs
integer, column numbers of housekeeping genes

method
method for the computation

na.rm
a logical value indicating whether NA values should be stripped before the computation proceeds.

Details

This function can be used to normalize real-time quantitative RT-PCR data. The default method "Vandesompele" was proposed by Vandesompele et al. (2002).

Currently, only the method by Vandesompele et al. (2002) is implemented.

Value

Normalized expression data

Author(s)

Dr. Matthias Kohl (SIRS-Lab GmbH) kohl@sirs-lab.com

References

Jo Vandesompele, Katleen De Preter, Filip Pattyn et al. (2002). Accurate normalization of real-time quantitative RT-PCR data by geometric averaging of multiple internal control genes. Genome Biology 2002. 3(7):research0034.1-0034.11. http://genomebiology.com/2002/3/7/research/0034/

Examples

data(SLqPCRdata)
relData <- apply(SLqPCRdata, 2, relQuantPCR)
geneStabM(relData[,c(3,4)])
exprData <- normPCR(SLqPCRdata, c(3,4))

[Package SLqPCR version 1.60.0 Index]