relQuantPCR {SLqPCR}	R Documentation

Compute relative expression values for realtime quantitative RT-PCR data

Description

Compute relative expression values for realtime quantitative RT-PCR data based on Ct or take-off values, respectively. The computations use the PCR efficiency.

Usage

relQuantPCR(x, E = 2, na.rm = FALSE)

Arguments

x
numeric vector containing raw data

E
PCR efficiency

na.rm
a logical value indicating whether NA values should be stripped before the computation proceeds.

Value

vector of relative expression values w.r.t. specified PCR efficiency.

Author(s)

Dr. Matthias Kohl (SIRS-Lab GmbH) kohl@sirs-lab.com

References

Jo Vandesompele, Katleen De Preter, Filip Pattyn et al. (2002). Accurate normalization of real-time quantitative RT-PCR data by geometric averaging of multiple internal control genes. Genome Biology 2002. 3(7):research0034.1-0034.11. http://genomebiology.com/2002/3/7/research/0034/

[Package SLqPCR version 1.60.0 Index]