selectHKgenes {SLqPCR}	R Documentation

Selection of reference/housekeeping genes

Description

This function can be used to determine a set of reference/housekeeping (HK) genes for gene expression experiments.

Usage

selectHKgenes(relData, method = "Vandesompele", minNrHK = 2, geneSymbol, 
	      trace = TRUE, na.rm = FALSE)

Arguments

relData
matrix or data.frame containing relative expression values

method
method to compute most stable genes

minNrHK
minimum number of HK genes that should be considered

geneSymbol
gene symbols

trace
logical, print additional information

na.rm
a logical value indicating whether NA values should be stripped before the computation proceeds.

Details

This function can be used to determine a set of reference/housekeeping (HK) genes for gene expression experiments. The default method "Vandesompele" was proposed by Vandesompele et al. (2002).

Currently, only the method by Vandesompele et al. (2002) is implemented.

Vandesompele et al. (2002) propose a cut-off value of 0.15 for the pairwise variation. Below this value the inclusion of an additional housekeeping gene is not required.

Value

If method = "Vandesompele" a list with the following components is returnd

ranking
ranking of genes from best to worst where the two most stable genes cannot be ranked

variation
pairwise variation during stepwise selection

meanM
average expression stability M

Author(s)

Dr. Matthias Kohl (SIRS-Lab GmbH) kohl@sirs-lab.com

References

Jo Vandesompele, Katleen De Preter, Filip Pattyn et al. (2002). Accurate normalization of real-time quantitative RT-PCR data by geometric averaging of multiple internal control genes. Genome Biology 2002. 3(7):research0034.1-0034.11. http://genomebiology.com/2002/3/7/research/0034/

Examples

data(vandesompele)
res.BM <- selectHKgenes(vandesompele[1:9,], method = "Vandesompele", geneSymbol = names(vandesompele), minNrHK = 2, trace = TRUE, na.rm = TRUE)

[Package SLqPCR version 1.60.0 Index]