vandesompele {SLqPCR}	R Documentation

Data set of Vandesompele et al (2002)

Description

This data set was used in Vandesompele et al (2002) to demonstrate normalization of real-time quantitative RT-PCR data by geometric averaging of housekeeping genes.

Usage

data(vandesompele)

Format

A data frame with 85 observations on the following 10 variables which stand for expression data of ten commonly used housekeeping genes

ACTB

actin, beta

B2M

beta-2-microglobulin

GAPD

glyceraldehyde-3-phosphate dehydrogenase

HMBS

hydroxymethylbilane synthase

HPRT1

hypoxanthine phosphoribosyltransferase 1

RPL13A

ribosomal protein L13a

SDHA

succinate dehydrogenase complex subunit A

TBP

TATA box binding protein

UBC

ubiquitin C

YWHAZ

tyrosine 3-monooxygenase/tryptophan 5-monooxygenase activation protein, zeta polypeptide

Details

The row names of this data set indicate the various human tissues which were investigated.

BM

9 normal bone-marrow samples

POOL

9 normal human tissues from pooled organs (heart, brain, fetal brain, lung, trachea, kidney, mammary gland, small intestine and uterus)

FIB

20 short-term cultured normal fibroblast samples from different individuals

LEU

13 normal leukocyte samples

NB

34 neuroblastoma cell lines (independently prepared in different labs from different patients)

Source

The data set was obtained from http://genomebiology.com/content/supplementary/gb-2002-3-7-research0034-s1.txt

References

Jo Vandesompele, Katleen De Preter, Filip Pattyn et al. (2002). Accurate normalization of real-time quantitative RT-PCR data by geometric averaging of multiple internal control genes. Genome Biology 2002. 3(7):research0034.1-0034.11. http://genomebiology.com/2002/3/7/research/0034/

Examples

data(vandesompele)
str(vandesompele)
rownames(vandesompele)

[Package SLqPCR version 1.60.0 Index]